Include usage information
All of the README files now have simple usage information in them that
says what you should do with the files if you want to use them, in
case someone is not fully versed in how Sublime works.

More work needs to be done here, particularly inside of the plugins
directory.