scratch_buffer command can now prompt for syntax
This modifies the scratch_buffer command so that if no explicit syntax
is provided to the command, it will prompt you with a quick panel to
select the syntax to use.

In use this collects all tmLanguage and sublime-syntax files and
displays them in sorted order. If a syntax is presented as both a
tmLanguage and sublime-syntax file, the sublime-syntax version takes
precedence in the list.

This is because versions of sublime that understand sublime-syntax
prefer that version of the syntax even if you specify the other
variant explicitly.

This presumes that someone is using a version of Sublime that handles
the newer syntax format, however (shame on you if you're not).

The supplied sublime-commands file also includes a new entry that
allows you to easily invoke the command in this mode.