Include new plugin set_vc_vars.py
This is a simple plugin example written as an answer on Stack Overflow
that demonstrates executing a batch file at plugin load time and
capturing the environment changes that it makes. The changes are then
applied to the environment of the currently running Sublime.

With this in place you can execute visual studio command line tools
from within Sublime build systems without having to launch Sublime
from a developer command prompt or run the batch file first every
time.

This could be easily modified to run any generic batch file at startup
for similar purposes. It's also currently set up to be Windows only,
although it could easily be modified to be cross platform.