Include new plugin wrap_text.py
This is a simple plugin example written as an answer on Stack Overflow
that demonstrates how to reflow multiline text to wrap at a certain column.
It is useful for Python docstrings.
The built in wrap paragraph functionality doesn't work the way I would expect,
and I can't find a use for it, so this essentially replaces that,
although it currently requires a selection rather than automatically working on the
paragraph at the caret.
Currently it only supports spaces for indentation, so there is some future work
required there to handle tabs.