Include toggle_setting_ext.py
This is an enhanced version of the toggle_setting command that allows
you to toggle arbitrary settings between arbitrary values, since the
internal toggle_setting command can only toggle settings that use
boolean values.

This is a modified version of the plugin from an SO answer that's made
more general to work with any setting. This tries to handle all error
states gracefully.

In particular, if the setting doesn't have a value yet, the first one
in the provided options is used. If the setting has a value that's not
in the list, the first option provided is also used, to bring the
setting in line for future toggles. Otherwise, the next setting in the
list provided is used.

This supports an arbitrary size of lists, but doesn't validate that
the list has at least two options in it under the theory that someone
will notice while they're making their key binding that they're
trying to bind a key that does nothing.