Update wrap_text.py to support Tab indentation
This involves checking the settings on the view to see which indentation style is being used - tabs or spaces - and how wide a single level of indentation is. (i.e. 4 spaces or one tab that acts like 4 spaces visually)
Additionally, the way the `textwrap` works is that a tab counts as a single character. This has the effect of wrapping too late - so we need to pretend the max wrap width is smaller (the difference between 1 character and the visual tab width being used in ST, multiplied by the indentation level).