Fix bug in wrap_text.py when there are no rulers
In some circumstances, the `rulers` setting on the view is not set at all, in other circumstances, it can be an empty array.
The previous code was failing when the array was empty, because it was trying to access index 0 for the default that only applied when the setting wasn't set at all.
To fix it, it will now substitute no value for an empty array, and if the array is empty, it will use the default of 72.