Update wrap_text.py to support reformatting line comments
(Block comments are already supported in that the start token is at the beginning of the text being formatted and the end token is... at the end - so there's no need to worry about rewrapping it.)
This is pretty complicated, because we need to be able to add and remove the line comment token in the right places. To achieve this, we are reusing the built in toggle comment implementation from the Default package to remove the line comments before we reflow/wrap the text. We then make use of the indentation feature of the text wrapper to reapply the line comments on every line. This way they all end up in the right place!
(Why didn't you just toggle the line comments back on afterwards, you ask? Well, then we'd have to adjust the max width we pass to the text wrapper so that it wouldn't wrap lines too long, and this approach seemed easier to understand. Plus, the selected text length could get affected that way.)