Include indent_test plugin and build
This includes a new prototype plugin for performing tests on indent
rules similar to how syntax testing works.

In use indent test files are created with a known name prefix and a
header line that marks the syntax that they use so that the
appropriate indent rules can be provided.

The provided build system will trigger automatically if a build is
invoked from within an indent test file, and also includes a variant
for running all indent tests all at once.

The test consists of inserting an unindented version of the test file
into a buffer, then running the indent command to indent the buffer,
which is then compared with the original to see if the indent worked
or not.