Use new command palette input for scratch syntax
This enhances the scratch buffer sample plugin so that when invoked
with a syntax of None from the command palette, the selection of the
syntax is done using the new command palette input mechanism.

The previous handling for this situation (prompting for the syntax in
a quick panel) is still done whenever the command is executed with a
syntax of None in all other cases, since the new handling only works
from within the command palette itself.

The plugin code adapts to this situation by using the input handler as
a simple proxy to get the appropriate information for the quick panel
to avoid code duplication.

Additionally, guard code is in place so that when the plugin is loaded
by versions of Sublime prior to build 3154 (where the command input
was first added) things gracefully fall back as expected.