Some slight cleanup in the scratch files plugin
This is a bit of README and code cleanup in the scratch files plugin,
along with a change to the way the command palette input works to take
advantage of changes made to this new input method after the initial
dev build that introduced it.

When the feature was initially added, the name of the command from the
command palette entry was used as the initial text in the command
palette. A subsequent release added the new `input_description` method
on command, which is queried to provide the text that's displayed in
lieu of the command palette entry.

As a consequence, initially the command palette did not tell you that
the scratch buffer command was going to prompt you for a syntax so
that it would look cleaner when you ran the command. Now it tells you
that it is going to prompt you, and the command tells you what you
should be entering instead, which looks cleaner.

As a part of the cleanup, the default is now to prompt you for a
syntax when you don't explicitly set one, rather than the previous
default of assuming plain text.