update scratch_files plugin to read syntax name from the YAML
This is a small tweak to improve the way the syntaxes are displayed in the scratch buffer prompt. Previously, it was just using the base name of the file, but in some cases, the base name and the name used by Sublime don't match, and it was confusing trying to find some syntaxes in the list.
This was especially noticable for files with lowercase names, as of course the items are sorted lexographically, making such syntaxes appear under the rest, seemingly "out of order". Whereas the "name" defined in the syntax definition itself is generally a more uniform, Title Case form.
To avoid requiring a YAML parser as a dependency, as there isn't one bundled with ST/ST's Python by default, it just uses a simple regex lookup.
This should probably also be updated in future to read the name of `.tmLanguage` files too...