Add preview for package and syntax name
This tweaks the syntax list input handler so that it shows a preview
pane that outlines the package that the syntax is from and the name
of the file that's providing it. This is a useful contextual clue at
times when you don't know where a syntax is coming from.

Also a bit of code tidying in a couple of spots to use a different
style of string formatting and remove some redundant path
manipulation.