Fix example key bindings for comment extension
This adds a couple of fixes to the example key bindings in the plugin
code that demonstrate how you could use the command to make your
comments continue on the next line.

There was an issue with line comment extension if the last line of the
buffer was a line comment that ends in a newline (that is, if visibly
the second to last line is a line comment and the cursor is at the
start  of the last line itself).

When that is the case, the scope for the line comment "bleeds" onto
the next line (I would guess because it's attached to the newline
character?), making the key binding think that it's on a line comment
and trying to extend it, which fails with an exception being thrown
because the current line is in fact not a line comment.

There was a similar issue with block comments, although in this case
the issue is the opposite; since the scope is definitely a block
comment, it would be continued with a new line containing an asterisk.

It seems like better UX to just insert a newline in this case, since
it is most likely a user decision to not have a prefix on the previous
line and so we should not add one to the next.

One last issue is that the block scope "bleeds" out of the block
comment if the line ends in the block terminator and the cursor is to
the right of it, e.g. if the line was "/**/|". This makes the  binding
trigger and extend the comment when it's already closed.

There is also a small tweak bundled in that combines the line comment
and C# documentation comment bindings into a single one with an
appropriate scope, since they are otherwise identical.