Fix scratch_buffer quick panel fallback selection
The existing `scratch_buffer` command is broken in that when you are
prompted for a syntax via a quick panel, although you can select a
syntax, it throws an error instead of actually creating a view in
response.

The reason for this is that the code assumes a method on the command
class that does not exist. This method was factored out in f1e8bb4
when the code for it was inlined in the `run` method without also
fixing the call point.