Allow wrap_text to leave the selection at the end of the wrapped text (#15)
which is more consistent with the `wrap_lines` behavior that it replaces/enhances, and so it does this by default now. I've kept the old behavior available from keybindings etc. by adding a `keep_selection` argument to the command. This way, if people want to be able to visually see what got wrapped (for example, when starting with no selection) then they still can. I feel this is better than adding a preference because maybe users will want two different keybindings/behaviors for different circumstances.