Include session cleaning script
This includes a version of the script I use locally to remove defunct
projects from Sublime so that they're not offered in the project quick
switch panel.

In use it loads up the `Session.sublime-session` file and removes any
workspace entries that reference files that don't seem to exist any
longer. Backups are kept of the session files as well.

The script attempts to determine the appropriate data directory based
on the platform you run it on and also allows you to specify a data
directory specifically on the command line as well.