Refactor the script a little bit
This refactors the code a little bit to fix some style/design issues
that remained from the original version of the script, which was much
more crude.

In particular the command line handling is cleaned up a bit so that
the session cleaning script gets told the data directory instead of
looking it up.

There is also a slight change to the code that collates the lists of
existing and missing projects to be more Pythonic.

Some work is still needed to clean up the exception handling and to
get things working on Windows, where the paths stored in the session
are not in a format that Python can internally process.