More refactoring and cleanup
This refactors the remaining code to be more clear and clean with
regards to exception handling.

Here the operations to load, save and clean the session file are split
apart and handle their own errors, making it much easier to see where
things might go wrong and ensure that the proper error messages are
triggered.

This also creates an abstraction for checking if a workspace is valid
or not, since on Windows more work needs to be done on the path in
order to check, given that Sublime uses a non-native file name format
for these entries in the session.