Implement proper existence checking on Windows
This implements proper checking for the existence of Sublime workspace
files on the Windows platform, where Sublime stores files in a
different format than one might expect in some cases.

In the case of a UNC file name to the workspace, the file is stored as
one might expect except with POSIX separators, for example:

    //weyr/Public/TestProject/TestProject.sublime-workspace

Such a path works with the `os.path.isfile()` method as one might
expect (including when the host name is a numeric IP) although if the
host is not available the method predictably returns `False`, which
may or may not be desirable.

In contrast, file names that normally have a drive letter (either a
local drive or due to a drive mapping) are stored with the drive being
considered a path segment off of the root instead of a normal drive.
e.g.:

    /C/Users/micro/Desktop/TestProject/TestProject.sublime-workspace

In this case the existence check fails because the file name is not in
an appropriate format.

This change modifies the handling specifically for windows where the
file is not a UNC path (as determined by checking the prefix) and
rewriting the file to include a drive letter.