Improve command line handling using argparse
This alters the script to use the `argparse` module to parse the
command line, which allows us to more naturally specify the data
directory.

In addition, a `--dry-run` parameter is also implemented, which makes
the session cleaner stop prior to actually rewriting the session file.
This allows you to see what it would do without actually taking the
action.

As a result of these changes, the README is also updated.