Document that core commands are not covered
The commands that are implemented directly in the Sublime core are not
handled by the general plugin host environment in the same way  as
package defined commands and thus cannot be introspected using the
mechanism that this plugin uses.

This change alters the documentation in the code and the entry in the
README to point this out, as before both indicated that all commands
were covered.