Include package bootstrapping example
This includes sample code for a Dependency (or a regular package, if
you wanted to) that is capable of automatically taking part of itself
to create a whole new `sublime-package` or keep that file updated as
the dependency is updated.

In a dependency, this allows the dependency to contain Sublime
commands and resources that can be made available as a regular package
even if/when Package Control installs the dependency in the new `lib`
folder of the most recent Subliem Text 3 builds.

The example shows:

  * Masking resources so Sublime won't load them except in the
    bootstrapped package.
  * Checking for the existence of a package
  * Creating or updating a package based on version checks
  * Ensuring that the update did not stop half way through
  * Taking care of resources that temporarily go away doing an update
    (e.g. syntax definitions that might be in use).

The code in the folder isn't laid out the way that an actual
dependency would be laid out, since the structure depends on the
contents of the resource in question (and it would make for an extra
`all` directory that just seems more confusing than helpful).