Documentation fixes and clarifications
This includes some missing information in the original commit that
better describes how things work in the sample.

In particular a description of what the bootstrapped package name
represents and how to change it was missing from the original version
of the code

Apart from this, various inconsistencies in word wrapping and spacing
have been resolved as well.