Include log_toggler.py
This includes a simple plugin that allows you to easily toggle the state
of the various logging mechanisms in Sublime on and off using a key
binding, command palette entry or menu item.

This also serves as a simple demonstration of using the `is_checked` and
`description` methods of the command classes to customize how a command
appears in the menu.

As there is no API for determining the current state of the logging
flags, the command tracks the current state itself, which means that it
will get out of sync with reality if you use some external mechanism to
toggle the logging.

This will clear itself up the next time you invoke the command (although
it may appear to do nothing), but on the whole this is probably not that
important since you likely already know what you're doing in that case.