Update the session cleaner to support Merge also
This modifies the existing session cleaning script so that it can clean
up the list of recent repositories in Sublime Merge similarly to how it
can clean up entries for defunct projects in Sublime Text.

The structure of the code remains largely the same, with slight tweaks
to the logic in some key locations based on the program being cleaned.

For backwards compatibility the default for the new `--program` command
line argument is `text`.