Handle newer Sublime Merge session formats
At some point (I haven't investigated exactly when), the Sublime Merge
session format diverged slightly from that used in Sublime Text in that
the list of recent items changed from a list of paths to a list of
objects that describe repositories.

This allows it to track not only the path but also things like the
window size, what commits are collapsed or expanded, and so on.

This small fix handles this by using a combination of the program whose
session is to be cleaned and an introspection of the items in the recent
list to determine if it needs to do more work to get the appropriate
path or not.

This allows it to work with both versions of the session format for
Merge.