Update menu_exec arguments
Instead of the menu_exec command  taking an argument that tells it to
invert the state of the show_panel_on_build setting temporarily, we
instead take an argument that specifies if the panel should be shown or
not, and not specifying it (or specifying `None`/`null`) honors the
global preference instead.

The logic here would be that if you want to execute a program from a
key binding or menu entry, you either explicitly do or do not want to
see the output, or you want to fall back on the global preference.