Clean up the open_all_found_files command
Minor cosmetic changes to the code for this command have been done,
mostly changing code style. In addition we now also disable the command
outside of Find in Files output so that the command can be included in
the command palette without constantly being available.

The scope used in `find_by_selector` now technically has a redundant
suffix since the command will only invoke in a file of the proper type,
but that has been left in place here.