Enhance command line handling
This augments the command line handling of the program a bit to include
some new options that we're going to need. In particular, there are now
options for specifying if workspaces/git repositories, folders and files
should be cleaned up, with the ability to select any of them.

Previously the default was inherently to do workspace/repo cleanup since
that was the only thing that the script did, but with the inclusion of
new items that has been removed in favor of requiring an explicit
declaration of what to clean up.

The idea is that with multiple items, a person should be free to clean
only what they want to (say if recent folders are on  currently
unmounted network shares). Although we could keep the old default, that
would require options to toggle the default off, and seems like bad UX
to make someone add in things they want and then maybe also have to
remove a thing they don't.