Update the sample script file
This updates the sample script file to include new options for what to
clean up when Sublime is launched. In addition this also includes the
notion (which is easily disable-able if desired) that running the script
while Sublime is running will create a new window, which makes taskbar
buttons for this nicer to use.

We also now find bash without hard coding it, for use on systems that
might have custom bash versions installed in non-standard locations.