Fix the package path for JSON in scratch_buffer
This fixes #26 reported by @Ultra-Instinct-05 ; as outlined in the issue
the JSON syntax changed from being in the `JavaScript` package to being
in its own package, which breaks the command palette entry for this
syntax.

Using `Prompt for Syntax` works as expected.

At some point, this plugin should be factored a bit to take advantage of
the new syntax endpoints added to the API in the 4k builds of Sublime.