Plugin for adding current file path to the sidebar
This plugin is the result of a question asked on StackOverflow and
provides a command which can be invoked to add the path of the current
file to the side bar, if it's not already there.

By request of the original question asker, the plugin also supports
Sublime Text 2 (which does not have an API for adjusting project data in
a window) by invoking Sublime on itself to add to the folder list.