Rework conditional logic to work in all cases
This one was a bit of a brain burner, but the context will now properly
work whether match_all is True or False, and whether the actual thing
you're comparing for is True or False.

This is done by using any() or all() depending on match_all, and also
remembering that for both functions, they search for True.

Thus, the state of the booleans in the comparison list needs to be True
when the desired comparison direction is reversed, and also True when
the desired comparison direction is forward.

It's actually wether or not we check that final value against True or
False that matters, in the end.